Expose all SecurityContext details via getters #885
